[PECOFF] Data type of SectionNumber is now unsigned (r203986).
So we don't need static_cast's to convert it from signed to unsigned.

llvm-svn: 203992

+++ lld/lib/ReaderWriter/PECOFF/ReaderCOFF.cpp
@@ -477,8 +477,7 @@ error_code FileCOFF::createDefinedSymbols(const SymbolVectorT &symbols,
       continue;
 
     const coff_section *sec;
-    if (error_code ec =
-            _obj->getSection(static_cast<uint16_t>(sym->SectionNumber), sec))
+    if (error_code ec = _obj->getSection(sym->SectionNumber, sec))
       return ec;
     assert(sec && "SectionIndex > 0, Sec must be non-null!");
 
@@ -527,8 +526,7 @@ error_code FileCOFF::cacheSectionAttributes() {
       continue;
 
     const coff_section *sec;
-    if (error_code ec =
-            _obj->getSection(static_cast<uint16_t>(sym->SectionNumber), sec))
+    if (error_code ec = _obj->getSection(sym->SectionNumber, sec))
       return ec;
 
     if (_merge.count(sec))
